requirements.txt
Django==4.0.4

django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-filter==22.1
django-import-export==2.8.0
django-cors-headers==3.12.0
requests==2.28.1